refactor Java unit tests into multiple files.
